refactor(react): extract voice audio subsystems
Split the 793-line use-voice-agent.ts by lifting the two default browser
Web Audio subsystems (createBrowserMicrophone/createBrowserSpeaker) plus
their structural DOM typings, PCM helpers (blockRms/toPcm16), and audio
contract types into a sibling voice-audio.ts. The hook keeps only its WS
transport + React-state logic; the audio graph stays isolated behind the
existing createMicrophone/createSpeaker injection seams and mockable in a
non-browser test env.
